README
## CODE2015 Hackathon - My Canadian Lifestyle

This is our submission for the CODE2015 Hackathon! Users are able to input their personal lifestyle preferences and the app will utilize open data in order to best determine which city in Canada will suite that individual best. We believe that healthy living depends heavily on the environment you’re in, and therefore it is crucial to determine where in Canada suites your personality best.

[Visit the application here!](http://mycanadianlifestyle.azurewebsites.net/)

### Our Team:

[John Vidal!](https://www.linkedin.com/in/jlvidal)
[Johnny Ji](http://johnnyji.com/)
[Elle Kasai](http://ellekasai.com/)
[Mariana D’Ornellas](https://www.linkedin.com/in/mdornellas)